feat: support auto-generated primary keys with IDENTITY columns (#503)
* feat: support auto-generated primary keys with IDENTITY columns

Adds support for auto-generated primary keys using IDENTITY columns.
Standard primary key columns in Entity Framework are translated to

The methods added in SpannerModelBuilderExtensions can be used to
configure the strategy for generating primary keys:

`SpannerModelBuilderExtensions.DisableIdentityColumns()` disables the use
of IDENTITY for primary keys.

`SpannerModelBuilderExtensions.UseAutoIncrement()` configures the Entity
Framework provider to generate `AUTO_INCREMENT` columns instead of a
a `default_sequence_kind` option set.

The `SpannerModelExtensions.SetIdentityOptions` method can be used to
manually configure the generation strategy. This can be used to set a
custom sequence kind.

* chore: cleanup

+++ Google.Cloud.EntityFrameworkCore.Spanner.IntegrationTests/MigrationTests/MigrationModels/AllColType.cs
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Google.Cloud.EntityFrameworkCore.Spanner.Storage;
 using Google.Cloud.Spanner.V1;
 using System;
 using System.Collections.Generic;
@@ -23,7 +22,7 @@ namespace Google.Cloud.EntityFrameworkCore.Spanner.IntegrationTests
 {
     public partial class AllColType
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public short? ColShort { get; set; }
         public int? ColInt { get; set; }
         public long? ColLong { get; set; }
